Various fixes for midiproc

+++ src/MIDIContainer.cpp
@@ -510,7 +510,7 @@ void MIDIContainer::SerializeAsStream(size_t subSongIndex, std::vector<MIDIStrea
 
         // Select which track can provide the next event.
         {
-            uint32_t NextTimestamp = ~0UL;
+            uint32_t NextTimestamp = ((uint32_t)~0UL);
 
             for (size_t i = 0; i < TrackCount; ++i)
             {
@@ -524,7 +524,7 @@ void MIDIContainer::SerializeAsStream(size_t subSongIndex, std::vector<MIDIStrea
                 }
             }
 
-            if (NextTimestamp == ~0UL)
+            if (NextTimestamp == ((uint32_t)~0UL))
                 break;
         }
 
@@ -905,10 +905,10 @@ uint32_t MIDIContainer::GetLoopBeginTimestamp(size_t subSongIndex, bool ms /* = 
     if (!ms)
         return Timestamp;
 
-    if (Timestamp != ~0UL)
+    if (Timestamp != ((uint32_t)~0UL))
         return TimestampToMS(Timestamp, TrackIndex);
 
-    return ~0UL;
+    return ((uint32_t)~0UL);
 }
 
 uint32_t MIDIContainer::GetLoopEndTimestamp(size_t subSongIndex, bool ms /* = false */) const
@@ -926,10 +926,10 @@ uint32_t MIDIContainer::GetLoopEndTimestamp(size_t subSongIndex, bool ms /* = fa
     if (!ms)
         return Timestamp;
 
-    if (Timestamp != ~0UL)
+    if (Timestamp != ((uint32_t)~0UL))
         return TimestampToMS(Timestamp, TrackIndex);
 
-    return ~0UL;
+    return ((uint32_t)~0UL);
 }
 
 void MIDIContainer::GetMetaData(size_t subSongIndex, MIDIMetaData & metaData)
@@ -1216,7 +1216,7 @@ void MIDIContainer::TrimStart()
 
 void MIDIContainer::TrimRange(size_t start, size_t end)
 {
-    uint32_t timestamp_first_note = ~0UL;
+    uint32_t timestamp_first_note = ((uint32_t)~0UL);
 
     for (size_t i = start; i <= end; ++i)
     {
@@ -1239,7 +1239,7 @@ void MIDIContainer::TrimRange(size_t start, size_t end)
         }
     }
 
-    if (timestamp_first_note < ~0UL && timestamp_first_note > 0)
+    if (timestamp_first_note < ((uint32_t)~0UL) && timestamp_first_note > 0)
     {
         for (size_t i = start; i <= end; ++i)
         {
